update user experience

+++ lib/stb_image_write.h
@@ -796,7 +796,7 @@ static void *stbiw__sbgrowf(void **arr, int increment, int itemsize) {
   void *p = STBIW_REALLOC_SIZED(
       *arr ? stbiw__sbraw(*arr) : 0,
       *arr ? (stbiw__sbm(*arr) * itemsize + sizeof(int) * 2) : 0,
-      itemsize * m + sizeof(int) * 2);
+      (unsigned long)itemsize * m + sizeof(int) * 2);
   STBIW_ASSERT(p);
   if (p) {
     if (!*arr) ((int *)p)[1] = 0;
@@ -1090,7 +1090,7 @@ static void stbiw__encode_png_line(unsigned char *pixels, int stride_bytes,
       stbi__flip_vertically_on_write ? -stride_bytes : stride_bytes;
 
   if (type == 0) {
-    memcpy(line_buffer, z, width * n);
+    memcpy(line_buffer, z, (size_t)width * n);
     return;
   }
 
@@ -1163,9 +1163,9 @@ STBIWDEF unsigned char *stbi_write_png_to_mem(const unsigned char *pixels,
     force_filter = -1;
   }
 
-  filt = (unsigned char *)STBIW_MALLOC((x * n + 1) * y);
+  filt = (unsigned char *)STBIW_MALLOC(((size_t)x * n + 1) * y);
   if (!filt) return 0;
-  line_buffer = (signed char *)STBIW_MALLOC(x * n);
+  line_buffer = (signed char *)STBIW_MALLOC((size_t)x * n);
   if (!line_buffer) {
     STBIW_FREE(filt);
     return 0;
@@ -1203,7 +1203,7 @@ STBIWDEF unsigned char *stbi_write_png_to_mem(const unsigned char *pixels,
     // when we get here, filter_type contains the filter type, and line_buffer
     // contains the data
     filt[j * (x * n + 1)] = (unsigned char)filter_type;
-    STBIW_MEMMOVE(filt + j * (x * n + 1) + 1, line_buffer, x * n);
+    STBIW_MEMMOVE(filt + j * (x * n + 1) + 1, line_buffer, (size_t)x * n);
   }
   STBIW_FREE(line_buffer);
   zlib = stbi_zlib_compress(filt, y * (x * n + 1), &zlen,
